==> Course_Materials_ALL/Course_Materials_Part2/Video_Lecture_NBs/NB_08_Importing CSV Files.ipynb <==
# Part 2: Full Data Workflow A-Z

## Importing CSV Files with pd.read_csv()

### First Steps

with open('titanic.csv') as f:
    text = f.readlines()

text

len(text)

print("Hello \nWorld!")

import pandas as pd

pd.read_csv("titanic.csv")

pd.read_csv("titanic.csv", index_col= "pclass")

pd.read_csv("titanic.csv", header = None)

pd.read_csv("titanic.csv", header = None, names= ["Alive", "Class", "Gender", "Age", "SipSp", "ParCh", "Fare", "Emb", "Deck"])

pd.read_csv("titanic.csv", header = 0, names= ["Alive", "Class", "Gender", "Age", "SipSp", "ParCh", "Fare", "Emb", "Deck"])

pd.read_csv("titanic.csv", usecols=["pclass", "survived", "sex", "age"])

titanic = pd.read_csv("titanic.csv", index_col= "pclass", usecols= ["pclass", "survived", "sex", "age"])

titanic.head()

titanic.columns = ["Alive", "Gender", "Age"]

titanic.index.name = "Class"

titanic.head()

### Importing messy CSV Files

import pandas as pd

pd.read_csv("titanic_raw.csv")

col_names = ["Survived", "Class", "Gender", "Age", "SipSp", "ParCh", "Fare", "Emb", "Deck"]

# titanic = pd.read_csv("titanic_raw.csv", skiprows= 3, skipfooter= 2, header = None, names = col_names) # old

titanic = pd.read_csv("titanic_raw.csv", skiprows= 3, skipfooter= 2, header = None, names = col_names, engine = "python") # new

titanic.head()

titanic.to_csv("titanic_imp.csv", index=False)

pd.read_csv("titanic_imp.csv")
